lambda/image-resizer/requirements.txt
Pillow==10.2.0
boto3>=1.26.0
